test: add an integration test to test tx-pool state cache

+++ test/src/node.rs
@@ -26,15 +26,20 @@ use std::process::{self, Child, Command, Stdio};
 use std::thread::sleep;
 use std::time::{Duration, Instant};
 
-struct ProcessGuard(pub Child);
+struct ProcessGuard {
+    pub child: Child,
+    pub killed: bool,
+}
 
 impl Drop for ProcessGuard {
     fn drop(&mut self) {
-        match self.0.kill() {
-            Err(e) => error!("Could not kill ckb process: {}", e),
-            Ok(_) => debug!("Successfully killed ckb process"),
+        if !self.killed {
+            match self.child.kill() {
+                Err(e) => error!("Could not kill ckb process: {}", e),
+                Ok(_) => debug!("Successfully killed ckb process"),
+            }
+            let _ = self.child.wait();
         }
-        let _ = self.0.wait();
     }
 }
 
@@ -532,12 +537,31 @@ impl Node {
             }
         };
 
-        self.guard = Some(ProcessGuard(child_process));
+        self.guard = Some(ProcessGuard {
+            child: child_process,
+            killed: false,
+        });
         self.node_id = Some(node_info.node_id);
     }
 
     pub fn stop(&mut self) {
         drop(self.guard.take())
+    }
+
+    #[cfg(target_os = "linux")]
+    pub fn stop_gracefully(&mut self) {
+        if let Some(mut guard) = self.guard.take() {
+            if !guard.killed {
+                // send SIGINT to the child
+                nix::sys::signal::kill(
+                    nix::unistd::Pid::from_raw(guard.child.id() as i32),
+                    nix::sys::signal::Signal::SIGINT,
+                )
+                .expect("cannot send ctrl-c");
+                let _ = guard.child.wait();
+                guard.killed = true;
+            }
+        }
     }
 
     pub fn export(&self, target: String) {
